refactor: move elderly school components to components/elderly-school/
